Fix #22 and #24 and add a failing test case for #23
- Do not mix `async` and `done` usage in jasmine env
- Stop polly instance on `test_skip`
- Add a test case for usage of global `Polly.register`
  (this currently fails on jest-circus)
- Update CI config to split test running in a better way
