+++ test/integration/jest-common/__recordings__/Polly-register_1263522794/setup-polly-jest-works-when-adapters-and-persisters-are-registered-with-Polly-register_638723674/recording.har
@@ -1,6 +1,6 @@
 {
   "log": {
-    "_recordingName": "setupPolly/should configure polly when needed",
+    "_recordingName": "Polly.register/setup-polly-jest works when adapters and persisters are registered with Polly.register",
     "creator": {
       "comment": "persister:fs",
       "name": "Polly.JS",
@@ -8,7 +8,7 @@
     },
     "entries": [
       {
-        "_id": "6051b8237686cd0ca6432c7b77d390fc",
+        "_id": "662efe612dbf05c41ce55a78acc5872f",
         "_order": 0,
         "cache": {},
         "request": {
@@ -40,32 +40,33 @@
               "value": "jsonplaceholder.typicode.com"
             }
           ],
-          "headersSize": 234,
+          "headersSize": 228,
           "httpVersion": "HTTP/1.1",
           "method": "GET",
           "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/-999999"
+          "url": "https://jsonplaceholder.typicode.com/posts/1"
         },
         "response": {
-          "bodySize": 2,
+          "bodySize": 408,
           "content": {
+            "_isBinary": true,
             "mimeType": "application/json; charset=utf-8",
-            "size": 2,
-            "text": "{}"
+            "size": 408,
+            "text": "[\"1f8b0800000000000003558f4b6a04410c43f7730a51eb6cb29d1be40ebd715c0a63e8fa8ccb1e2684dc3d744308012324901fe8eb02945cf4b75aae787d39a2fdd9b0d859ae282b7b4032f0214a279c93fb2e81e9e361953d3054852a61e0533923dd3066d838cace1b7ba55b9413fc3eeae7c1bda7091858b9d4a6c5d67f1d9c9a4b7a1542475fbc67f648079f93d5e27cd36c5bff87471b3b57981019c7dd53166284b4adf7b1c2b3c17928d739890de73a67031fecc680b8de2ca831cae5fb0715c693e824010000\"]"
           },
           "cookies": [],
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 02 Aug 2021 08:39:34 GMT"
+              "value": "Mon, 02 Aug 2021 08:39:28 GMT"
             },
             {
               "name": "content-type",
               "value": "application/json; charset=utf-8"
             },
             {
-              "name": "content-length",
-              "value": "2"
+              "name": "transfer-encoding",
+              "value": "chunked"
             },
             {
               "name": "connection",
@@ -81,11 +82,11 @@
             },
             {
               "name": "x-ratelimit-remaining",
-              "value": "998"
+              "value": "999"
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1627893596"
+              "value": "1623347254"
             },
             {
               "name": "vary",
@@ -113,7 +114,7 @@
             },
             {
               "name": "etag",
-              "value": "W/\"2-vyGp6PvFo4RvsFtPoIWeCReyIC8\""
+              "value": "W/\"124-yiKdLzqO5gfBrJFrcdJ8Yq0LGnU\""
             },
             {
               "name": "via",
@@ -121,7 +122,11 @@
             },
             {
               "name": "cf-cache-status",
-              "value": "MISS"
+              "value": "HIT"
+            },
+            {
+              "name": "age",
+              "value": "495"
             },
             {
               "name": "expect-ct",
@@ -129,7 +134,7 @@
             },
             {
               "name": "report-to",
-              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=08L4z7aOfC07Ed3dUVCfbiKoLEzbyzeTbw71EjRpfLxsmQA6yk1NhgGnOWaJalzwlOyGAzGC8MFvoOH564CAB5ExwtB5UOJnrQr6SVuAsBeZrfP5zsuNsu2ya%2FvUboNVguPzMj82Xb%2BnSIVwI%2FEw\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
+              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=A5L9rxMasjqnLLMx6ywEVlv16o0Tx4ZUanO5isn8UYrr08nEaCC1A9b8oJt5Wii5MYiVdDcPQQhg9YGz%2BmNBncLstw4CDjIWciqXSFRORdZGRI1OjIXQgC%2BSxu5lkuLrWXI0x5Wx7RuF9J1GiG3Q\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
             },
             {
               "name": "nel",
@@ -141,21 +146,25 @@
             },
             {
               "name": "cf-ray",
-              "value": "6785ff166d24d463-HAM"
+              "value": "6785fef62fb3d43f-HAM"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
             },
             {
               "name": "alt-svc",
               "value": "h3-27=\":443\"; ma=86400, h3-28=\":443\"; ma=86400, h3-29=\":443\"; ma=86400, h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1046,
+          "headersSize": 1088,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
-          "status": 404,
-          "statusText": "Not Found"
+          "status": 200,
+          "statusText": "OK"
         },
-        "startedDateTime": "2021-08-02T08:38:14.563Z",
-        "time": 318,
+        "startedDateTime": "2021-08-02T08:38:09.370Z",
+        "time": 108,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -163,7 +172,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 318
+          "wait": 108
         }
       }
     ],
