+++ test/integration/jest-common/__recordings__/setupPolly_466486458/should-have-polly-running-with-it_2136411752/recording.har
@@ -1,6 +1,6 @@
 {
   "log": {
-    "_recordingName": "setupPolly/should configure polly when needed",
+    "_recordingName": "setupPolly/should have polly running with `it`",
     "creator": {
       "comment": "persister:fs",
       "name": "Polly.JS",
@@ -8,7 +8,7 @@
     },
     "entries": [
       {
-        "_id": "6051b8237686cd0ca6432c7b77d390fc",
+        "_id": "32797aa7f35cbb21eb4061a6fd1fe0fd",
         "_order": 0,
         "cache": {},
         "request": {
@@ -40,32 +40,33 @@
               "value": "jsonplaceholder.typicode.com"
             }
           ],
-          "headersSize": 234,
+          "headersSize": 228,
           "httpVersion": "HTTP/1.1",
           "method": "GET",
           "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/-999999"
+          "url": "https://jsonplaceholder.typicode.com/posts/2"
         },
         "response": {
-          "bodySize": 2,
+          "bodySize": 400,
           "content": {
+            "_isBinary": true,
             "mimeType": "application/json; charset=utf-8",
-            "size": 2,
-            "text": "{}"
+            "size": 400,
+            "text": "[\"1f8b08000000000000032d4e316e03310cdbf30ac2739776cc0ffa875b7435db08f0c98e2407288afcbdf0b51b2992227f2e4099417fafe58ad7974575c1b713a66663b9a2dca782916004cb29edbd7e2f655d9d3e0f248fd19d78680a370bae50a8254c6fdae01cce1badd235517beb8e9d924250fe3803c6fbe4669ff34b25b137b1aaa91a78f436474a12a3bb773cd8fe1f1fbd315285986b8b7e28ad6a6cb606c8a0ab1cb06ea8dc3535307a841e33b00c67214c4361b3352997e72ffc7a199216010000\"]"
           },
           "cookies": [],
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 02 Aug 2021 08:39:34 GMT"
+              "value": "Mon, 02 Aug 2021 08:39:28 GMT"
             },
             {
               "name": "content-type",
               "value": "application/json; charset=utf-8"
             },
             {
-              "name": "content-length",
-              "value": "2"
+              "name": "transfer-encoding",
+              "value": "chunked"
             },
             {
               "name": "connection",
@@ -81,11 +82,11 @@
             },
             {
               "name": "x-ratelimit-remaining",
-              "value": "998"
+              "value": "999"
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1627893596"
+              "value": "1627498661"
             },
             {
               "name": "vary",
@@ -113,7 +114,7 @@
             },
             {
               "name": "etag",
-              "value": "W/\"2-vyGp6PvFo4RvsFtPoIWeCReyIC8\""
+              "value": "W/\"116-jnDuMpjju89+9j7e0BqkdFsVRjs\""
             },
             {
               "name": "via",
@@ -121,7 +122,11 @@
             },
             {
               "name": "cf-cache-status",
-              "value": "MISS"
+              "value": "HIT"
+            },
+            {
+              "name": "age",
+              "value": "22364"
             },
             {
               "name": "expect-ct",
@@ -129,7 +134,7 @@
             },
             {
               "name": "report-to",
-              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=08L4z7aOfC07Ed3dUVCfbiKoLEzbyzeTbw71EjRpfLxsmQA6yk1NhgGnOWaJalzwlOyGAzGC8MFvoOH564CAB5ExwtB5UOJnrQr6SVuAsBeZrfP5zsuNsu2ya%2FvUboNVguPzMj82Xb%2BnSIVwI%2FEw\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
+              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=brX%2FDn%2FHHZIrWhSssLe9BNQyBQf6A2uFUF5qHl0AXM6775YJuKpuBwiRQcntDJNHC%2BV9OkwhPttQAqk%2Bz97ff%2FfEMRNdZWd1akINBP3KJDFuE9p78VHOaHty5qNQbdHzWHVOln9m1r3zYv3tL%2Byg\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
             },
             {
               "name": "nel",
@@ -141,21 +146,25 @@
             },
             {
               "name": "cf-ray",
-              "value": "6785ff166d24d463-HAM"
+              "value": "6785fef36a5e2c36-FRA"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
             },
             {
               "name": "alt-svc",
               "value": "h3-27=\":443\"; ma=86400, h3-28=\":443\"; ma=86400, h3-29=\":443\"; ma=86400, h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1046,
+          "headersSize": 1098,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
-          "status": 404,
-          "statusText": "Not Found"
+          "status": 200,
+          "statusText": "OK"
         },
-        "startedDateTime": "2021-08-02T08:38:14.563Z",
-        "time": 318,
+        "startedDateTime": "2021-08-02T08:38:08.959Z",
+        "time": 78,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -163,7 +172,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 318
+          "wait": 78
         }
       }
     ],
