+++ test/integration/jest-common/__recordings__/setupPolly_466486458/should-work-with-each-capturing-recording-for-post-3_1208924099/recording.har
@@ -1,6 +1,6 @@
 {
   "log": {
-    "_recordingName": "setupPolly/should configure polly when needed",
+    "_recordingName": "setupPolly/should work with each, capturing recording for post #3",
     "creator": {
       "comment": "persister:fs",
       "name": "Polly.JS",
@@ -8,7 +8,7 @@
     },
     "entries": [
       {
-        "_id": "6051b8237686cd0ca6432c7b77d390fc",
+        "_id": "4da0ad841cbd2b68368a694196131a3f",
         "_order": 0,
         "cache": {},
         "request": {
@@ -40,32 +40,33 @@
               "value": "jsonplaceholder.typicode.com"
             }
           ],
-          "headersSize": 234,
+          "headersSize": 228,
           "httpVersion": "HTTP/1.1",
           "method": "GET",
           "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/-999999"
+          "url": "https://jsonplaceholder.typicode.com/posts/3"
         },
         "response": {
-          "bodySize": 2,
+          "bodySize": 402,
           "content": {
+            "_isBinary": true,
             "mimeType": "application/json; charset=utf-8",
-            "size": 2,
-            "text": "{}"
+            "size": 402,
+            "text": "[\"1f8b0800000000000003558e316ac4401004f37b45a3d88971763ff01f2e19ed366660b523edcc081be3bf1f2becc0593534457ddf80259de3bd2e77bcbecca913df2e0c8dc6e58e8582cd1a3d541c478a2bf8c9513424d43a370cee6c4d02472a7477816b403296cbb45afdba44014d0f83b3e24883e6e0a39fd6720f096eb052844542615b57079b7eb0579d2e48fdf7add66ce89a8e930d524abaf4d0dc66856397a112391efd2f9ed86d0c03351d56d5c04093d506279d6cbfd1b79f277355343e1b010000\"]"
           },
           "cookies": [],
           "headers": [
             {
               "name": "date",
-              "value": "Mon, 02 Aug 2021 08:39:34 GMT"
+              "value": "Mon, 02 Aug 2021 08:39:29 GMT"
             },
             {
               "name": "content-type",
               "value": "application/json; charset=utf-8"
             },
             {
-              "name": "content-length",
-              "value": "2"
+              "name": "transfer-encoding",
+              "value": "chunked"
             },
             {
               "name": "connection",
@@ -85,7 +86,7 @@
             },
             {
               "name": "x-ratelimit-reset",
-              "value": "1627893596"
+              "value": "1627892756"
             },
             {
               "name": "vary",
@@ -113,7 +114,7 @@
             },
             {
               "name": "etag",
-              "value": "W/\"2-vyGp6PvFo4RvsFtPoIWeCReyIC8\""
+              "value": "W/\"11b-USacuIw5a/iXAGdNKBvqr/TbMTc\""
             },
             {
               "name": "via",
@@ -121,7 +122,11 @@
             },
             {
               "name": "cf-cache-status",
-              "value": "MISS"
+              "value": "HIT"
+            },
+            {
+              "name": "age",
+              "value": "825"
             },
             {
               "name": "expect-ct",
@@ -129,7 +134,7 @@
             },
             {
               "name": "report-to",
-              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=08L4z7aOfC07Ed3dUVCfbiKoLEzbyzeTbw71EjRpfLxsmQA6yk1NhgGnOWaJalzwlOyGAzGC8MFvoOH564CAB5ExwtB5UOJnrQr6SVuAsBeZrfP5zsuNsu2ya%2FvUboNVguPzMj82Xb%2BnSIVwI%2FEw\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
+              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=cVFVyI0FETIeONKi%2F56%2BCIAo2GVVRU6Zt7nmeFCPam3cec2%2F6tjgJ9XGXnq3vHKU7sDqG00D%2FxqFKzEaAf5muhdW8Gdqm%2BMFpaIXxm0rwf2Cq3ULuNIRatbqmm4tdChlvRsadgPS8DORyZItMk3W\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
             },
             {
               "name": "nel",
@@ -141,21 +146,25 @@
             },
             {
               "name": "cf-ray",
-              "value": "6785ff166d24d463-HAM"
+              "value": "6785fef69d8b4175-HAM"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
             },
             {
               "name": "alt-svc",
               "value": "h3-27=\":443\"; ma=86400, h3-28=\":443\"; ma=86400, h3-29=\":443\"; ma=86400, h3=\":443\"; ma=86400"
             }
           ],
-          "headersSize": 1046,
+          "headersSize": 1094,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
-          "status": 404,
-          "statusText": "Not Found"
+          "status": 200,
+          "statusText": "OK"
         },
-        "startedDateTime": "2021-08-02T08:38:14.563Z",
-        "time": 318,
+        "startedDateTime": "2021-08-02T08:38:09.468Z",
+        "time": 64,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -163,7 +172,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 318
+          "wait": 64
         }
       }
     ],
