--- test/integration/jest/__recordings__/setupPolly_466486458/should-configure-polly-when-needed_241621399/recording.har
@@ -1,163 +1,0 @@
-{
-  "log": {
-    "_recordingName": "setupPolly/should configure polly when needed",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "2.6.2"
-    },
-    "entries": [
-      {
-        "_id": "6051b8237686cd0ca6432c7b77d390fc",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [
-            {
-              "_fromType": "array",
-              "name": "accept",
-              "value": "*/*"
-            },
-            {
-              "_fromType": "array",
-              "name": "user-agent",
-              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept-encoding",
-              "value": "gzip,deflate"
-            },
-            {
-              "_fromType": "array",
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "host",
-              "value": "jsonplaceholder.typicode.com"
-            }
-          ],
-          "headersSize": 234,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/-999999"
-        },
-        "response": {
-          "bodySize": 2,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 2,
-            "text": "{}"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2020-09-13T08:23:14.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "d2deecf723f3bd29325bd54d0399c24181568449394"
-            }
-          ],
-          "headers": [
-            {
-              "name": "date",
-              "value": "Sat, 14 Sep 2019 08:23:15 GMT"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "content-length",
-              "value": "2"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "_fromType": "array",
-              "name": "set-cookie",
-              "value": "__cfduid=d2deecf723f3bd29325bd54d0399c24181568449394; expires=Sun, 13-Sep-20 08:23:14 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 14 Sep 2019 12:23:15 GMT"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"2-vyGp6PvFo4RvsFtPoIWeCReyIC8\""
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "EXPIRED"
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "cf-ray",
-              "value": "5160f52dafcf2724-FRA"
-            }
-          ],
-          "headersSize": 723,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 404,
-          "statusText": "Not Found"
-        },
-        "startedDateTime": "2019-09-14T08:23:14.785Z",
-        "time": 308,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 308
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
