--- test/integration/jest/__recordings__/setupPolly_466486458/should-have-polly-running-with-it-_2136411752/recording.har
@@ -1,171 +1,0 @@
-{
-  "log": {
-    "_recordingName": "setupPolly/should have polly running with `it`",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "2.6.2"
-    },
-    "entries": [
-      {
-        "_id": "32797aa7f35cbb21eb4061a6fd1fe0fd",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [
-            {
-              "_fromType": "array",
-              "name": "accept",
-              "value": "*/*"
-            },
-            {
-              "_fromType": "array",
-              "name": "user-agent",
-              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept-encoding",
-              "value": "gzip,deflate"
-            },
-            {
-              "_fromType": "array",
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "host",
-              "value": "jsonplaceholder.typicode.com"
-            }
-          ],
-          "headersSize": 228,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/2"
-        },
-        "response": {
-          "bodySize": 400,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 400,
-            "text": "[\"1f8b08000000000000032d4e316e03310cdbf30ac2739776cc0ffa875b7435db08f0c98e2407288afcbdf0b51b2992227f2e4099417fafe58ad7974575c1b713a66663b9a2dca782916004cb29edbd7e2f655d9d3e0f248fd19d78680a370bae50a8254c6fdae01cce1badd235517beb8e9d924250fe3803c6fbe4669ff34b25b137b1aaa91a78f436474a12a3bb773cd8fe1f1fbd315285986b8b7e28ad6a6cb606c8a0ab1cb06ea8dc3535307a841e33b00c67214c4361b3352997e72ffc7a199216010000\"]"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2020-09-13T08:23:14.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "d19593185b1b56784da586947641d31c41568449394"
-            }
-          ],
-          "headers": [
-            {
-              "name": "date",
-              "value": "Sat, 14 Sep 2019 08:23:14 GMT"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "transfer-encoding",
-              "value": "chunked"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "_fromType": "array",
-              "name": "set-cookie",
-              "value": "__cfduid=d19593185b1b56784da586947641d31c41568449394; expires=Sun, 13-Sep-20 08:23:14 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 14 Sep 2019 12:23:14 GMT"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"116-jnDuMpjju89+9j7e0BqkdFsVRjs\""
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "HIT"
-            },
-            {
-              "name": "age",
-              "value": "3726"
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "cf-ray",
-              "value": "5160f52b5cb9d6e9-FRA"
-            },
-            {
-              "name": "content-encoding",
-              "value": "gzip"
-            }
-          ],
-          "headersSize": 765,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2019-09-14T08:23:14.408Z",
-        "time": 70,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 70
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
