--- test/integration/jest/__recordings__/setupPolly_466486458/should-work-with-each-capturing-recording-for-post-1_1242479337/recording.har
@@ -1,171 +1,0 @@
-{
-  "log": {
-    "_recordingName": "setupPolly/should work with each, capturing recording for post #1",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "2.6.2"
-    },
-    "entries": [
-      {
-        "_id": "662efe612dbf05c41ce55a78acc5872f",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [
-            {
-              "_fromType": "array",
-              "name": "accept",
-              "value": "*/*"
-            },
-            {
-              "_fromType": "array",
-              "name": "user-agent",
-              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept-encoding",
-              "value": "gzip,deflate"
-            },
-            {
-              "_fromType": "array",
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "host",
-              "value": "jsonplaceholder.typicode.com"
-            }
-          ],
-          "headersSize": 228,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/1"
-        },
-        "response": {
-          "bodySize": 408,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 408,
-            "text": "[\"1f8b0800000000000003558f4b6a04410c43f7730a51eb6cb29d1be40ebd715c0a63e8fa8ccb1e2684dc3d744308012324901fe8eb02945cf4b75aae787d39a2fdd9b0d859ae282b7b4032f0214a279c93fb2e81e9e361953d3054852a61e0533923dd3066d838cace1b7ba55b9413fc3eeae7c1bda7091858b9d4a6c5d67f1d9c9a4b7a1542475fbc67f648079f93d5e27cd36c5bff87471b3b57981019c7dd53166284b4adf7b1c2b3c17928d739890de73a67031fecc680b8de2ca831cae5fb0715c693e824010000\"]"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2020-09-13T08:23:14.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "d19d12545a9d36c2c045899b483fbbea71568449394"
-            }
-          ],
-          "headers": [
-            {
-              "name": "date",
-              "value": "Sat, 14 Sep 2019 08:23:14 GMT"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "transfer-encoding",
-              "value": "chunked"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "_fromType": "array",
-              "name": "set-cookie",
-              "value": "__cfduid=d19d12545a9d36c2c045899b483fbbea71568449394; expires=Sun, 13-Sep-20 08:23:14 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 14 Sep 2019 12:23:14 GMT"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"124-yiKdLzqO5gfBrJFrcdJ8Yq0LGnU\""
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "HIT"
-            },
-            {
-              "name": "age",
-              "value": "1256"
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "cf-ray",
-              "value": "5160f52bcfe9bf14-FRA"
-            },
-            {
-              "name": "content-encoding",
-              "value": "gzip"
-            }
-          ],
-          "headersSize": 765,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2019-09-14T08:23:14.484Z",
-        "time": 69,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 69
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
