--- test/integration/jest/__recordings__/setupPolly_466486458/should-work-with-each-capturing-recording-for-post-3_1208924099/recording.har
@@ -1,171 +1,0 @@
-{
-  "log": {
-    "_recordingName": "setupPolly/should work with each, capturing recording for post #3",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "2.6.2"
-    },
-    "entries": [
-      {
-        "_id": "4da0ad841cbd2b68368a694196131a3f",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [
-            {
-              "_fromType": "array",
-              "name": "accept",
-              "value": "*/*"
-            },
-            {
-              "_fromType": "array",
-              "name": "user-agent",
-              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept-encoding",
-              "value": "gzip,deflate"
-            },
-            {
-              "_fromType": "array",
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "host",
-              "value": "jsonplaceholder.typicode.com"
-            }
-          ],
-          "headersSize": 228,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/3"
-        },
-        "response": {
-          "bodySize": 402,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 402,
-            "text": "[\"1f8b0800000000000003558e316ac4401004f37b45a3d88971763ff01f2e19ed366660b523edcc081be3bf1f2becc0593534457ddf80259de3bd2e77bcbecca913df2e0c8dc6e58e8582cd1a3d541c478a2bf8c9513424d43a370cee6c4d02472a7477816b403296cbb45afdba44014d0f83b3e24883e6e0a39fd6720f096eb052844542615b57079b7eb0579d2e48fdf7add66ce89a8e930d524abaf4d0dc66856397a112391efd2f9ed86d0c03351d56d5c04093d506279d6cbfd1b79f277355343e1b010000\"]"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2020-09-13T08:23:14.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "df7b8194fd7fa9864c78f161d51d857991568449394"
-            }
-          ],
-          "headers": [
-            {
-              "name": "date",
-              "value": "Sat, 14 Sep 2019 08:23:14 GMT"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "transfer-encoding",
-              "value": "chunked"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "_fromType": "array",
-              "name": "set-cookie",
-              "value": "__cfduid=df7b8194fd7fa9864c78f161d51d857991568449394; expires=Sun, 13-Sep-20 08:23:14 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 14 Sep 2019 12:23:14 GMT"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"11b-USacuIw5a/iXAGdNKBvqr/TbMTc\""
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "HIT"
-            },
-            {
-              "name": "age",
-              "value": "5304"
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "cf-ray",
-              "value": "5160f52ccef8c281-FRA"
-            },
-            {
-              "name": "content-encoding",
-              "value": "gzip"
-            }
-          ],
-          "headersSize": 765,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2019-09-14T08:23:14.638Z",
-        "time": 69,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 69
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
