--- test/integration/jest/__recordings__/setupPolly_466486458/with-skipped-test_1941074839/should-record_255246347/recording.har
@@ -1,189 +1,0 @@
-{
-  "log": {
-    "_recordingName": "setupPolly/with skipped test/should record",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "5.0.0"
-    },
-    "entries": [
-      {
-        "_id": "662efe612dbf05c41ce55a78acc5872f",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [
-            {
-              "_fromType": "array",
-              "name": "accept",
-              "value": "*/*"
-            },
-            {
-              "_fromType": "array",
-              "name": "user-agent",
-              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept-encoding",
-              "value": "gzip,deflate"
-            },
-            {
-              "_fromType": "array",
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "host",
-              "value": "jsonplaceholder.typicode.com"
-            }
-          ],
-          "headersSize": 228,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/1"
-        },
-        "response": {
-          "bodySize": 408,
-          "content": {
-            "_isBinary": true,
-            "mimeType": "application/json; charset=utf-8",
-            "size": 408,
-            "text": "[\"1f8b0800000000000003558f4b6a04410c43f7730a51eb6cb29d1be40ebd715c0a63e8fa8ccb1e2684dc3d744308012324901fe8eb02945cf4b75aae787d39a2fdd9b0d859ae282b7b4032f0214a279c93fb2e81e9e361953d3054852a61e0533923dd3066d838cace1b7ba55b9413fc3eeae7c1bda7091858b9d4a6c5d67f1d9c9a4b7a1542475fbc67f648079f93d5e27cd36c5bff87471b3b57981019c7dd53166284b4adf7b1c2b3c17928d739890de73a67031fecc680b8de2ca831cae5fb0715c693e824010000\"]"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2020-10-07T09:32:26.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "sameSite": "Lax",
-              "value": "de30c714a345075fcc3915c4dcbd16a231599471146"
-            }
-          ],
-          "headers": [
-            {
-              "name": "date",
-              "value": "Mon, 07 Sep 2020 09:32:26 GMT"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "transfer-encoding",
-              "value": "chunked"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "_fromType": "array",
-              "name": "set-cookie",
-              "value": "__cfduid=de30c714a345075fcc3915c4dcbd16a231599471146; expires=Wed, 07-Oct-20 09:32:26 GMT; path=/; domain=.typicode.com; HttpOnly; SameSite=Lax"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            },
-            {
-              "name": "x-ratelimit-limit",
-              "value": "1000"
-            },
-            {
-              "name": "x-ratelimit-remaining",
-              "value": "997"
-            },
-            {
-              "name": "x-ratelimit-reset",
-              "value": "1598841974"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "max-age=43200"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "expires",
-              "value": "-1"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"124-yiKdLzqO5gfBrJFrcdJ8Yq0LGnU\""
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "HIT"
-            },
-            {
-              "name": "age",
-              "value": "13633"
-            },
-            {
-              "name": "cf-request-id",
-              "value": "050981eefa00000b3fcb0ea200000001"
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "cf-ray",
-              "value": "5cef6c2b2af10b3f-AMS"
-            },
-            {
-              "name": "content-encoding",
-              "value": "gzip"
-            }
-          ],
-          "headersSize": 878,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2020-09-07T09:32:25.221Z",
-        "time": 131,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 131
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
