+++ test/jest-circus-integration.config.json
@@ -1,5 +1,5 @@
 {
   "rootDir": "../",
-  "testMatch": ["<rootDir>/test/integration/jest/**/*.test.js"],
+  "testMatch": ["<rootDir>/test/integration/jest-{common,circus}/**/*.test.js"],
   "testRunner": "jest-circus/runner"
 }
