+++ test/jest-jasmine-integration.config.json
@@ -1,5 +1,5 @@
 {
   "rootDir": "../",
-  "testMatch": ["<rootDir>/test/integration/jest/**/*.test.js"],
+  "testMatch": ["<rootDir>/test/integration/jest-common/**/*.test.js"],
   "testRunner": "jasmine2"
 }
