--- test/unit/setup-polly-unsupported-env.test.js
@@ -1,11 +1,0 @@
-/**
- * @jest-environment jest-environment-node
- */
-import { describe, it, expect } from '@jest/globals';
-import { setupPolly } from '../..';
-
-describe('setupPolly in unsupported environment', () => {
-  it('throws when called', () => {
-    expect(() => setupPolly()).toThrow();
-  });
-});
